Refactored existing BundleService to an abstract class implementing the interface, so that it can be used in the service for localization
