+++ Nwazet.Commerce.csproj
@@ -204,7 +204,8 @@
     <Compile Include="Controllers\StripeController.cs" />
     <Compile Include="Drivers\AdvancedSKUProductPartDriver.cs" />
     <Compile Include="Drivers\AdvancedSKUsSiteSettingPartDriver.cs" />
-    <Compile Include="Drivers\AttributeLocalizationProductAttributesPartDriver.cs" /> 
+    <Compile Include="Drivers\AttributeLocalizationProductAttributesPartDriver.cs" />
+    <Compile Include="Drivers\BundleLocalizationPartDriver.cs" />
     <Compile Include="Drivers\BundlePartDriver.cs" />
     <Compile Include="Drivers\ECommerceCurrencySiteSettingsPartDriver.cs" />
     <Compile Include="Drivers\OrderPartDriver.cs" />
@@ -231,7 +232,7 @@
     <Compile Include="Handlers\AttributePartHandler.cs" />
     <Compile Include="Handlers\AttributesPartHandler.cs" />
     <Compile Include="Handlers\ECommerceCurrencySiteSettingsPartHandler.cs" />
-    <Compile Include="Handlers\LocalizationAttributePartHandler.cs" /> 
+    <Compile Include="Handlers\LocalizationAttributePartHandler.cs" />
     <Compile Include="Handlers\OrderPartHandler.cs" />
     <Compile Include="Handlers\ProductSettingsPartHandler.cs" />
     <Compile Include="Handlers\StateOrCountryTaxPartHandler.cs" />
@@ -251,7 +252,7 @@
     <Compile Include="Menus\PromotionAdminMenu.cs" />
     <Compile Include="Menus\ProductAdminMenu.cs" />
     <Compile Include="Menus\ShippingAdminMenu.cs" />
-    <Compile Include="Migrations\AttributesLocalizationExtensionsMigrations.cs" /> 
+    <Compile Include="Migrations\AttributesLocalizationExtensionsMigrations.cs" />
     <Compile Include="Migrations\BundleMigrations.cs" />
     <Compile Include="Migrations\BundlesLocalizationExtensionsMigrations.cs" />
     <Compile Include="Migrations\CommerceMigrations.cs" />
@@ -331,16 +332,18 @@
     <Compile Include="Routes\StripeRoutes.cs" />
     <Compile Include="Routes\OrderRoutes.cs" />
     <Compile Include="Routes\CommerceRoutes.cs" />
+    <Compile Include="Services\BundleProductLocalizationBudleService.cs" />
+    <Compile Include="Services\BundleServiceBase.cs" />
     <Compile Include="Services\CurrencyProviderBase.cs" />
     <Compile Include="Services\IBundleService.cs" />
-    <Compile Include="Services\IProductAttributeAdminServices.cs" /> 
+    <Compile Include="Services\IProductAttributeAdminServices.cs" />
     <Compile Include="Services\IProductAttributeLocalizationServices.cs" />
     <Compile Include="Services\ISKUGenerationServices.cs" />
     <Compile Include="Services\ISKUUniquenessHelper.cs" />
-    <Compile Include="Services\LatestProductAttributeAdminServices.cs" /> 
+    <Compile Include="Services\LatestProductAttributeAdminServices.cs" />
     <Compile Include="Services\LocalizationSKUUniquenessHelper.cs" />
     <Compile Include="Services\IProductAttributeNameService.cs" />
-    <Compile Include="Services\ProductAttributeAdminServices.cs" /> 
+    <Compile Include="Services\ProductAttributeAdminServices.cs" />
     <Compile Include="Services\ProductAttributeLocalizationServices.cs" />
     <Compile Include="Services\ProductAttributeNameService.cs" />
     <Compile Include="Services\Currency.cs" />
@@ -398,8 +401,8 @@
     <Compile Include="Services\UnitedStates.cs" />
     <Compile Include="Settings\BundleProductLocalizationEditorEvents.cs" />
     <Compile Include="Settings\BundleProductLocalizationSettings.cs" />
-    <Compile Include="Settings\ProductAttributeLocalizationEditorEvents.cs" /> 
-    <Compile Include="Settings\ProductAttributeLocalizationSettings.cs" /> 
+    <Compile Include="Settings\ProductAttributeLocalizationEditorEvents.cs" />
+    <Compile Include="Settings\ProductAttributeLocalizationSettings.cs" />
     <Compile Include="Shapes\ProductAttributesLocalizationAdminShapeProvider.cs" />
     <Compile Include="Tokens\CartTokens.cs" />
     <Compile Include="Tokens\OrderTokens.cs" />
@@ -408,7 +411,7 @@
     <Compile Include="ViewModels\AdvancedSKUProductEditorViewModel.cs" />
     <Compile Include="ViewModels\ECommerceCurrencySiteSettingsViewModel.cs" />
     <Compile Include="ViewModels\ECommerceSettingsViewModel.cs" />
-    <Compile Include="ViewModels\AttributeLocalizationProductAttributesPartEditViewModel.cs" /> 
+    <Compile Include="ViewModels\AttributeLocalizationProductAttributesPartEditViewModel.cs" />
     <Compile Include="ViewModels\ProductAttributePartDisplayViewModel.cs" />
     <Compile Include="ViewModels\ProductAttributePartEditViewModel.cs" />
     <Compile Include="ViewModels\AttributesIndexViewModel.cs" />
@@ -530,10 +533,10 @@
     <None Include="Views\ProductAttributesAdminList.cshtml" />
     <None Include="Views\ProductAttributesAdminListItem.cshtml" />
     <None Include="Views\DefinitionTemplates\BundleProductLocalizationSettings.cshtml" />
-    <None Include="Views\DefinitionTemplates\ProductAttributeLocalizationSettings.cshtml" /> 
+    <None Include="Views\DefinitionTemplates\ProductAttributeLocalizationSettings.cshtml" />
     <None Include="Views\ECommerceSettingsAdmin\Index.cshtml" />
     <None Include="Views\EditorTemplates\Parts\AdvancedSKUProduct.cshtml" />
-    <None Include="Views\EditorTemplates\Parts\AttributeLocalizationProductAttributes.cshtml" /> 
+    <None Include="Views\EditorTemplates\Parts\AttributeLocalizationProductAttributes.cshtml" />
     <None Include="Views\EditorTemplates\SiteSettings\AdvancedSKUs.cshtml" />
     <None Include="Views\EditorTemplates\SiteSettings\Currency.cshtml" />
     <None Include="Views\ProductAttributesLocalizationAdminListItem.cshtml" />
@@ -570,4 +573,4 @@
     <PostBuildEvent>
     </PostBuildEvent>
   </PropertyGroup>
-</Project> 
+</Project>
